Move {read,write}Geometry{From,To}File to GeoLib/IO.
+ Update the namespace from FileIO to IO.
+ move Legacy/OGSIOVer4* to GeoLib/IO.
